meme-ai-backend/requirements.txt
accelerate==1.12.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fastapi==0.122.0
filelock==3.20.0
firebase_admin==7.1.0
fsspec==2025.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6
numpy==2.3.5
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
